Generate our template-pages via a layout template.
This will make updating our layout to use bootstrap much
more simple, and updates #15.

We've added a new Makefile target to generate the templates,
fixed our deployment script to invoke it, and renamed all the
files.

The template page `login.template` was incorrectly named (`template`
suffix rather than `tmpl`), so that was fixed.