Simplified DNS lookup, by avoiding the Amazon API.
This commit updates #14, by changing the way we lookup the
current value(s) of the given names.  Rather than fetching
the live data from the Amazon API we lookup via DNS.

If a user is registered and has `foo.dhcp.io` we do a
lookup of the name `foo.dhcp.io`, and rely on the caching
present in the local resolver.

This is more efficient than fetching the contents of the complete
`dhcp.io` zone, then parsing each record to see if it matches
`foo`.

TODO:

* Update the API we bundle, such that we can use `upsert`.

Upsert allows us to send a message which says:

* Set foo.dhcp.io (A) -> 1.2.3.4

Rather than having to :

* Download the zone.
* Find the current value.
    * Delete the value, explicitly.
    * Add the value with the new IP.

The whole thing we're trying to avoid is having to download
the zone, because internally the Amazon API allows you to only
do that in stages:

* Download the first 50 records.
   * Download the next 50 records.
   * Download the next 50 records.
   * Are we done?  If not download the next records.
   * etc.